Added: - Quality fixes for function arguments and * imports - Doctrings and usage commands - Scripts tested locally with example commands from old vision.py - vision.py modified to point user to new script - updation of README - updation of example commands in recipes
